# api/docs/swagger.yaml
basePath: /todo
definitions:
  main.Todo:
    properties:
      created_at:
        type: string
      desc:
        type: string
      done:
        type: boolean
      done_at:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample API for managing todos.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo API
  version: "1.0"
paths:
  /todo:
    get:
      description: Retrieve a list of all Todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Todo'
            type: array
      summary: Get all Todos
      tags:
      - Todos
  /todo/create:
    post:
      consumes:
      - application/json
      description: Create a new Todo with provided information
      parameters:
      - description: Todo Information
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/main.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Todo'
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Create a new Todo
      tags:
      - Todos
  /todo/delete/{id}:
    delete:
      description: Delete a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Todo deleted successfully
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Delete a Todo
      tags:
      - Todos
  /todo/getuser/{id}:
    get:
      description: Retrieve details of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Get a Todo by ID
      tags:
      - Todos
  /todo/update-status/{id}:
    patch:
      description: Update the status of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Update Todo Status
      tags:
      - Todos
  /todo/update/{id}:
    patch:
      consumes:
      - application/json
      description: Update details of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Todo Information
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/main.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Update a Todo
      tags:
      - Todos
swagger: "2.0"